Stop writing blocks at the transition block
Include account state of the transition block, but not the block or transactions themselves.

This is due to an off-by-one error in the PoS migration tool, which migrates the history only until the block before the transition block (as the end of a range is exclusive), but takes the account state at the transition block.